CR: simplify integration lookup
